update model & add schema and delete response
