+++ src/model.rs
@@ -1,35 +1,16 @@
-use chrono::prelude::*;
 use serde::{Deserialize, Serialize};
-use std::sync::Arc;
-use tokio::sync::Mutex;
+use sqlx::FromRow;
+use uuid::Uuid;
 
+#[derive(Debug, FromRow, Deserialize, Serialize)]
 #[allow(non_snake_case)]
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct ToDo {
-    pub id: Option<String>,
+pub struct ToDoModel {
+    pub id: Uuid,
     pub title: String,
     pub content: String,
     pub complete: Option<bool>,
-    pub createdAt: Option<DateTime<Utc>>,
-    pub updatedAt: Option<DateTime<Utc>>,
-}
-
-pub type DB = Arc<Mutex<Vec<ToDo>>>;
-
-pub fn todo_db() -> DB {
-    Arc::new(Mutex::new(Vec::new()))
-}
-
-#[derive(Debug, Deserialize, Default)]
-pub struct QueryOptions {
-    pub page: Option<usize>,
-    pub limit: Option<usize>,
-}
-
-#[allow(non_snake_case)]
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct UpdateToDo {
-    pub title: Option<String>,
-    pub content: Option<String>,
-    pub complete: Option<bool>,
+    #[serde(rename = "createdAt")]
+    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
+    #[serde(rename = "updatedAt")]
+    pub updated_at: Option<chrono::DateTime<chrono::Utc>>,
 }
